+++ src/actions/ReceiveDropdown.tsx
@@ -7,6 +7,7 @@ import { FlashNotification } from '../components/navigation/FlashNotification'
 import { Airship, showError } from '../components/services/AirshipInstance'
 import { lstrings } from '../locales/strings'
 import { getExchangeDenom, selectDisplayDenom } from '../selectors/DenominationSelectors'
+import { useSelector } from '../types/reactRedux'
 import { ThunkAction } from '../types/reduxTypes'
 import { NavigationBase } from '../types/routerTypes'
 import { calculateSpamThreshold, convertNativeToDisplay, zeroString } from '../util/utils'
@@ -28,14 +29,14 @@ export function showReceiveDropdown(navigation: NavigationBase, transaction: Edg
     const wallet = account.currencyWallets[walletId]
     if (wallet == null) return
 
-    const { fiatCurrencyCode } = wallet
+    const isoFiatCurrencyCode = useSelector(state => state.ui.settings.defaultIsoFiat)
 
     // Never stack dropdowns:
     if (receiveDropdownShowing) return
 
     // Check the spam limits:
     const { spamFilterOn } = state.ui.settings
-    const exchangeRate = state.exchangeRates[`${currencyCode}_${fiatCurrencyCode}`]
+    const exchangeRate = state.exchangeRates[`${currencyCode}_${isoFiatCurrencyCode}`]
     const exchangeDenom = getExchangeDenom(wallet.currencyConfig, tokenId)
     const spamThreshold = calculateSpamThreshold(exchangeRate, exchangeDenom)
     if (spamFilterOn && (zeroString(exchangeRate) || lt(nativeAmount, spamThreshold))) {