+++ src/components/scenes/Fio/FioStakingOverviewScene.tsx
@@ -6,7 +6,7 @@ import { sprintf } from 'sprintf-js'
 
 import { refreshAllFioAddresses } from '../../../actions/FioAddressActions'
 import fioLogo from '../../../assets/images/fio/fio_logo.png'
-import { getSymbolFromCurrency } from '../../../constants/WalletAndCurrencyConstants'
+import { getFiatSymbol } from '../../../constants/WalletAndCurrencyConstants'
 import { useAsyncEffect } from '../../../hooks/useAsyncEffect'
 import { useWatch } from '../../../hooks/useWatch'
 import { formatNumber, formatTimeDate } from '../../../locales/intl'
@@ -179,6 +179,7 @@ export const FioStakingOverviewScene = connect<StateProps, DispatchProps, OwnPro
       }
     } = ownProps
     const { account } = state.core
+    const { defaultIsoFiat, defaultFiat } = state.ui.settings
     const currencyWallet = account.currencyWallets[walletId]
     const currencyCode = getCurrencyCode(currencyWallet, tokenId)
 
@@ -190,7 +191,7 @@ export const FioStakingOverviewScene = connect<StateProps, DispatchProps, OwnPro
 
     const defaultDenomination = getExchangeDenomByCurrencyCode(currencyWallet.currencyConfig, currencyCode)
     const stakingDefaultCryptoAmount = convertNativeToDenomination(defaultDenomination.multiplier)(stakedNativeAmount ?? '0')
-    const stakingFiatBalance = convertCurrency(state, currencyCode, currencyWallet.fiatCurrencyCode, stakingDefaultCryptoAmount)
+    const stakingFiatBalance = convertCurrency(state, currencyCode, defaultIsoFiat, stakingDefaultCryptoAmount)
     const stakingFiatBalanceFormat = formatNumber(stakingFiatBalance && gt(stakingFiatBalance, '0.000001') ? stakingFiatBalance : 0, { toFixed: 2 })
 
     return {
@@ -198,8 +199,8 @@ export const FioStakingOverviewScene = connect<StateProps, DispatchProps, OwnPro
       stakingCryptoAmountFormat,
       stakingFiatBalanceFormat,
       currencyDenomination,
-      fiatCurrencyCode: currencyWallet.fiatCurrencyCode.replace('iso:', ''),
-      fiatSymbol: getSymbolFromCurrency(currencyWallet.fiatCurrencyCode)
+      fiatCurrencyCode: defaultFiat,
+      fiatSymbol: getFiatSymbol(defaultIsoFiat)
     }
   },
   dispatch => ({