+++ src/components/ui4/BalanceCardUi4.tsx
@@ -3,13 +3,13 @@ import { View } from 'react-native'
 import IonIcon from 'react-native-vector-icons/Ionicons'
 
 import { toggleAccountBalanceVisibility } from '../../actions/LocalSettingsActions'
-import { getSymbolFromCurrency } from '../../constants/WalletAndCurrencyConstants'
+import { getFiatSymbol } from '../../constants/WalletAndCurrencyConstants'
 import { useHandler } from '../../hooks/useHandler'
 import { formatNumber } from '../../locales/intl'
 import { lstrings } from '../../locales/strings'
 import { useDispatch, useSelector } from '../../types/reactRedux'
 import { NavigationBase } from '../../types/routerTypes'
-import { getTotalFiatAmountFromExchangeRates } from '../../util/utils'
+import { getTotalFiatAmountFromExchangeRates, removeIsoPrefix } from '../../util/utils'
 import { AnimatedNumber } from '../common/AnimatedNumber'
 import { EdgeTouchableOpacity } from '../common/EdgeTouchableOpacity'
 import { TransferModal } from '../modals/TransferModal'
@@ -43,8 +43,8 @@ export const BalanceCardUi4 = (props: Props) => {
   const fiatAmount = useSelector(state => getTotalFiatAmountFromExchangeRates(state, defaultIsoFiat))
   const exchangeRates = useSelector(state => state.exchangeRates)
 
-  const fiatSymbol = defaultIsoFiat ? getSymbolFromCurrency(defaultIsoFiat) : ''
-  const fiatCurrencyCode = defaultIsoFiat.replace('iso:', '')
+  const fiatSymbol = defaultIsoFiat ? getFiatSymbol(defaultIsoFiat) : ''
+  const fiatCurrencyCode = removeIsoPrefix(defaultIsoFiat)
   const formattedFiat = isBalanceVisible ? formatNumber(fiatAmount, { toFixed: 2 }) : lstrings.redacted_placeholder
 
   const exchangeRatesReady = exchangeRates != null && Object.keys(exchangeRates).length > 0