+++ src/plugins/gui/fiatPlugin.tsx
@@ -44,6 +44,7 @@ export const SendErrorBackPressed = 'SendErrorBackPressed'
 export const executePlugin = async (params: {
   account: EdgeAccount
   disklet: Disklet
+  defaultIsoFiat: string
   deviceId: string
   direction: 'buy' | 'sell'
   disablePlugins?: NestedDisableMap
@@ -60,6 +61,7 @@ export const executePlugin = async (params: {
   const {
     disablePlugins = {},
     account,
+    defaultIsoFiat,
     deviceId,
     direction,
     disklet,
@@ -285,6 +287,7 @@ export const executePlugin = async (params: {
   const paymentTypes = paymentType != null ? [paymentType] : []
   const startPluginParams: FiatPluginStartParams = {
     direction,
+    defaultIsoFiat,
     regionCode,
     paymentTypes,
     forceFiatCurrencyCode,