+++ src/selectors/SettingsSelectors.ts
@@ -1,6 +1,7 @@
 import { RootState } from '../types/reduxTypes'
+import { removeIsoPrefix } from '../util/utils'
 
 export const getDefaultFiat = (state: RootState) => {
   const defaultIsoFiat: string = state.ui.settings.defaultIsoFiat
-  return defaultIsoFiat.replace('iso:', '')
+  return removeIsoPrefix(defaultIsoFiat)
 }